Add files via upload
This project will discuss about some the applications of Hugging Face Transformers in Natural Language Processing (NLP).

Hugging Face Transformers package is very popular and versatile Python library that provides pre-trained models for a variety of applications in NLP, as well other areas such as image analysis, audio analysis, multimodal analysis (optical character recognition, video classification, visual question answering and many more). Some of the exercises included in this project are : Sentiment Analysis, Topic Classification, Text Generator: Masked Language Modeling, Name Entity Recognition, Question Answering, Text Summarization, and Translation. 

Mastering NLP with transformers provides a competitive advantage in an increasingly competitive job market for sure. We can open new careers in the fields of AI and NLP which are predicted to continue to grow. Let's continue to improve our data analysis capabilities and develop applications that are intelligent and have real impact.

Studying NLP and Transformers is an investment for the future so this technology continues to develop and shape a future where human and machine interactions are increasingly sophisticated and we must be part of the innovation that will change that world.